agreements datacube documentation

agreements datacube documentation

Usage

agreements

Format

The agreements datacube is a list that contains the following 7 datasets: DESTA, GPTAD, LABPTA, TOTA, TREND, HUGGO, AIGGO. For more information and references to each of the datasets used, please use the manydata::call_sources() and manydata::compare_dimensions() functions.

DESTA:

A dataset with 959 observations and the following 10 variables: manyID, treatyID, Title, Begin, AgreementType, DocType, GeogArea, Signature, Force, destaID.

GPTAD:

A dataset with 340 observations and the following 10 variables: manyID, treatyID, Title, Begin, AgreementType, DocType, GeogArea, Signature, Force, gptadID.

LABPTA:

A dataset with 483 observations and the following 7 variables: manyID, treatyID, Title, Begin, Signature, Force, labptaID.

TOTA:

A dataset with 442 observations and the following 7 variables: manyID, treatyID, Title, Begin, Signature, Force, totaID.

TREND:

A dataset with 710 observations and the following 7 variables: manyID, treatyID, Title, Begin, Signature, Force, trendID.

HUGGO:

A dataset with 610 observations and the following 11 variables: manyID, treatyID, Title, Begin, Signature, Force, End, url, Citation, TreatyText, Coder.

AIGGO:

A dataset with 1389 observations and the following 8 variables: manyID, treatyID, Title, Begin, Signature, Force, accessionC, accessionP.

Source

DESTA: A. Dür, L. Baccini, and M. Elsig. “The Design of International Trade Agreements: Introducing a NewDatabase”. In: The Review of International Organizations 9.3 (2014), pp. 353-375.

GPTAD: W. B. Group. Global Preferential Trade Agreement Database (GPTAD). Online database. publisher: WorldBank Group. 2014. https://wits.worldbank.org/gptad/library.aspx.

LABPTA: D. Raess, A. Dür, and D. Sari. “Protecting labor rights in preferential trade agreements: The role of trade unions, left governments, and skilled labor”. In: The Review of International Organizations 2.13(2018), pp. 143-162. DOI: 10.1007/s11558-018-9301-z.

TOTA: W. Alschner, J. Seiermann, and D. Skougarevskiy. Text-as-data analysis of preferential tradeagreements: Mapping the PTA landscape. 2017. https://github.com/mappingtreaties/tota.git.

TREND: J. Morin, A. Dür, and L. Lechner. “Mapping the trade and environment nexus: Insights from a newdataset”. In: Global Environmental Politics 18.1 (2018), pp. 122-139.

HUGGO: J. Hollway. Trade agreements for manydata. 2021.

AIGGO: J. Hollway. Trade agreements for manydata. 2021.

Details

#> $DESTA
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             959|     0     |       0       |
#> |   treatyID    |  character  |             959|     0     |       0       |
#> |     Title     |  character  |             959|     0     |       0       |
#> |     Begin     |    mdate    |             959|    81     |     8.45      |
#> | AgreementType |  character  |             959|     0     |       0       |
#> |    DocType    |  character  |             959|     0     |       0       |
#> |   GeogArea    |  character  |             959|     0     |       0       |
#> |   Signature   |    mdate    |             959|    81     |     8.45      |
#> |     Force     |    mdate    |             959|    242    |     25.23     |
#> |    destaID    |  character  |             959|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $GPTAD
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             340|     0     |       0       |
#> |   treatyID    |  character  |             340|     0     |       0       |
#> |     Title     |  character  |             340|     0     |       0       |
#> |     Begin     |    mdate    |             340|    30     |     8.82      |
#> | AgreementType |  character  |             340|     0     |       0       |
#> |    DocType    |  character  |             340|     0     |       0       |
#> |   GeogArea    |  character  |             340|    75     |     22.06     |
#> |   Signature   |    mdate    |             340|    53     |     15.59     |
#> |     Force     |    mdate    |             340|    63     |     18.53     |
#> |    gptadID    |  character  |             340|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $LABPTA
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             483|     0     |       0       |
#> |   treatyID    |  character  |             483|     0     |       0       |
#> |     Title     |  character  |             483|     0     |       0       |
#> |     Begin     |    mdate    |             483|     0     |       0       |
#> |   Signature   |    mdate    |             483|     0     |       0       |
#> |     Force     |    mdate    |             483|     0     |       0       |
#> |   labptaID    |  character  |             483|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TOTA
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             442|     0     |       0       |
#> |   treatyID    |  character  |             442|     0     |       0       |
#> |     Title     |  character  |             442|     0     |       0       |
#> |     Begin     |    mdate    |             442|     0     |       0       |
#> |   Signature   |    mdate    |             442|     0     |       0       |
#> |     Force     |    mdate    |             442|     3     |     0.68      |
#> |    totaID     |  character  |             442|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TREND
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             710|     0     |       0       |
#> |   treatyID    |  character  |             710|     0     |       0       |
#> |     Title     |  character  |             710|     0     |       0       |
#> |     Begin     |    mdate    |             710|     5     |      0.7      |
#> |   Signature   |    mdate    |             710|     5     |      0.7      |
#> |     Force     |    mdate    |             710|     5     |      0.7      |
#> |    trendID    |  character  |             710|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             611|     0     |       0       |
#> |     Title     |  character  |             611|     0     |       0       |
#> |     Begin     |    mdate    |             611|     0     |       0       |
#> |   Signature   |    mdate    |             611|     0     |       0       |
#> |     Force     |    mdate    |             611|    59     |     9.66      |
#> |      End      |    mdate    |             611|    113    |     18.49     |
#> |      url      |  character  |             611|     7     |     1.15      |
#> |   Citation    |  character  |             611|    610    |     99.84     |
#> |  TreatyText   |  character  |             611|     0     |       0       |
#> |   treatyID    |  character  |             611|     0     |       0       |
#> |     Coder     |  character  |             611|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $AIGGO
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            1389|     0     |       0       |
#> |   treatyID    |  character  |            1389|     0     |       0       |
#> |     Title     |  character  |            1389|     0     |       0       |
#> |     Begin     |    mdate    |            1389|    115    |     8.28      |
#> |   Signature   |    mdate    |            1389|    137    |     9.86      |
#> |     Force     |    mdate    |            1389|    252    |     18.14     |
#> |  accessionC   |  character  |            1389|   1000    |     71.99     |
#> |  accessionP   |  character  |            1389|   1139    |      82       |
#> ----------------------------------------------------------------------------

URL

DESTA: https://www.designoftradeagreements.org/downloads/

GPTAD: https://wits.worldbank.org/gptad/library.aspx

LABPTA: https://doi.org/10.1007/s11558-018-9301-z

TOTA: https://github.com/mappingtreaties/tota.git

TREND: http://www.chaire-epi.ulaval.ca/en/trend

HUGGO: Hand-coded data by the GGO team

AIGGO: Machine-generated data by the GGO team

Mapping

DESTA: Variable Mapping

from	to

base_treaty	destaID

name	Title

year	Signature

entryforceyear	Force

typememb	DocType

entry_type	AgreementType

regioncon	GeogArea

GPTAD: Variable Mapping

from	to

Common.Name	Title

Date.of.Signature	Signature

Date.of.Entry.into.Force	Force

Type	DocType

Type	AgreementType

Type	GeogArea

LABPTA: Variable Mapping

from	to

Number	labptaID

Name	Title

year	Signature

year	Force

TOTA: Variable Mapping

from	to

name	Title

date_signed	Signature

date_into_force	Force

TREND: Variable Mapping

from	to

Trade.Agreement	Title

Year	Signature

Year	Force