memberships datacube documentation

memberships datacube documentation

Usage

memberships

Format

The memberships datacube is a list that contains the following 3 datasets: GPTAD_MEM, DESTA_MEM, HUGGO_MEM. For more information and references to each of the datasets used, please use the manydata::call_sources() and manydata::compare_dimensions() functions.

GPTAD_MEM:

A dataset with 2192 observations and the following 9 variables: manyID, treatyID, Title, Begin, stateID, Signature, Force, StateName, gptadID.

DESTA_MEM:

A dataset with 7466 observations and the following 9 variables: manyID, treatyID, Title, Begin, stateID, Signature, Force, StateName, destaID.

HUGGO_MEM:

A dataset with 5608 observations and the following 20 variables: manyID, treatyID, Title, Begin, stateID, Signature, Force, End, StateName, gptadID, destaID, StateSignature, StateRatification, StateForce, StateEnd, Rat=Notif, Coder, Source, Succession, Accession.

Source

GPTAD_MEM: W. B. Group. Global Preferential Trade Agreement Database (GPTAD). https://wits.worldbank.org/gptad/library.aspx. 2013.

DESTA_MEM: A. Dür, L. Baccini, and M. Elsig. "The Design of International Trade Agreements: Introducing a NewDatabase". In: The Review of International Organizations 9.3 (2014), pp. 353-375.

HUGGO_MEM: J. Hollway. Trade agreements for manydata. 2021.

Details

#> $GPTAD_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2192|     0     |       0       |
#> |   treatyID    |  character  |            2192|     0     |       0       |
#> |     Title     |  character  |            2192|     0     |       0       |
#> |     Begin     |    mdate    |            2192|    99     |     4.52      |
#> |    stateID    |  character  |            2192|    55     |     2.51      |
#> |   Signature   |    mdate    |            2192|    310    |     14.14     |
#> |     Force     |    mdate    |            2192|    285    |      13       |
#> |   StateName   |  character  |            2192|    36     |     1.64      |
#> |    gptadID    |  character  |            2192|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $DESTA_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            7466|     0     |       0       |
#> |   treatyID    |  character  |            7466|     0     |       0       |
#> |     Title     |  character  |            7466|     0     |       0       |
#> |     Begin     |    mdate    |            7466|    630    |     8.44      |
#> |    stateID    |  character  |            7466|    11     |     0.15      |
#> |   Signature   |    mdate    |            7466|    630    |     8.44      |
#> |     Force     |    mdate    |            7466|   2864    |     38.36     |
#> |   StateName   |  character  |            7466|     0     |       0       |
#> |    destaID    |  character  |            7466|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO_MEM
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |            5608|     1     |     0.02      |
#> |    treatyID     |  character  |            5608|     1     |     0.02      |
#> |      Title      |  character  |            5608|     1     |     0.02      |
#> |      Begin      |    mdate    |            5608|     1     |     0.02      |
#> |     stateID     |  character  |            5608|     2     |     0.04      |
#> |    Signature    |    mdate    |            5608|     3     |     0.05      |
#> |      Force      |    mdate    |            5608|    367    |     6.54      |
#> |       End       |    mdate    |            5608|    556    |     9.91      |
#> |    StateName    |  character  |            5608|     1     |     0.02      |
#> |     gptadID     |  character  |            5608|   3436    |     61.27     |
#> |     destaID     |  character  |            5608|    553    |     9.86      |
#> | StateSignature  |    mdate    |            5608|    903    |     16.1      |
#> |StateRatification|    mdate    |            5608|   4893    |     87.25     |
#> |   StateForce    |    mdate    |            5608|   1125    |     20.06     |
#> |    StateEnd     |    mdate    |            5608|    630    |     11.23     |
#> |    Rat=Notif    |  character  |            5608|   5536    |     98.72     |
#> |      Coder      |  character  |            5608|    16     |     0.29      |
#> |     Source      |  character  |            5608|   4886    |     87.13     |
#> |   Succession    |  character  |            5608|     1     |     0.02      |
#> |    Accession    |    mdate    |            5608|   4679    |     83.43     |
#> ------------------------------------------------------------------------------

URL

GPTAD_MEM: https://wits.worldbank.org/gptad/library.aspx

DESTA_MEM: https://www.designoftradeagreements.org/downloads/

HUGGO_MEM: Hand-coded data by the GGO team.

Mapping

GPTAD_MEM: Variable Mapping

from	to

Membership	StateName

Common.Name	Title

Date.of.Signature	Signature

Date.of.Entry.into.Force	Force

DESTA_MEM: Variable Mapping

from	to

base_treaty	destaID

name	Title

year	Signature

entryforceyear	Force

c1:c91	stateID