memberships database documentation

memberships database documentation

Usage

memberships

Format

The memberships database is a list that contains the following 2 datasets: GPTAD_MEM, DESTA_MEM. For more information and references to each of the datasets used, please use the data_source() and data_contrast() functions.

GPTAD_MEM:

A dataset with 2198 observations and the following 9 variables: manyID, stateID, Title, Beg, Signature, Force, StateName, gptadID, treatyID.

DESTA_MEM:

A dataset with 7492 observations and the following 9 variables: manyID, stateID, Title, Beg, Signature, Force, StateName, destaID, treatyID.

Details

## $GPTAD_MEM
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
## ----------------------------------------------------------------------------
## |    manyID     |  character  |            2198|     0     |       0       |
## |    stateID    |  character  |            2198|    130    |     5.91      |
## |     Title     |  character  |            2198|     0     |       0       |
## |      Beg      |    mdate    |            2198|    99     |      4.5      |
## |   Signature   |    mdate    |            2198|    328    |     14.92     |
## |     Force     |    mdate    |            2198|    283    |     12.88     |
## |   StateName   |  character  |            2198|     0     |       0       |
## |    gptadID    |  character  |            2198|     0     |       0       |
## |   treatyID    |  character  |            2198|     0     |       0       |
## ----------------------------------------------------------------------------
## 
## 
## $DESTA_MEM
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
## ----------------------------------------------------------------------------
## |    manyID     |  character  |            7492|     0     |       0       |
## |    stateID    |  character  |            7492|    10     |     0.13      |
## |     Title     |  character  |            7492|     0     |       0       |
## |      Beg      |    mdate    |            7492|    630    |     8.41      |
## |   Signature   |    mdate    |            7492|    630    |     8.41      |
## |     Force     |    mdate    |            7492|   2886    |     38.52     |
## |   StateName   |  character  |            7492|     0     |       0       |
## |    destaID    |  character  |            7492|     0     |       0       |
## |   treatyID    |  character  |            7492|     0     |       0       |
## ----------------------------------------------------------------------------